loss+or_profit
Write a program to calculate the loss or profit (i.e., difference on price) in a book
purchase at outlet or online. The program gets the price of the book you bought at outlet
as input. Then it asks for the average price of the book, if bought online. The program
computes the difference between these prices and displays the output.